Refactor dz2
