+++ dz2/SummAndUmnojenie.java
@@ -1,3 +1,5 @@
+package dz2;
+
 import java.util.Scanner;
 
 public class SummAndUmnojenie {
