+++ dz2/dz2.java
@@ -1,3 +1,5 @@
+package dz2;
+
 public class dz2 {
     
 }
