+++ dz2/exponentiation.java
@@ -1,3 +1,5 @@
+package dz2;
+
 import java.util.Scanner;
 
 public class exponentiation {
