+++ dz2/maximum_of_three_numbers.java
@@ -1,4 +1,4 @@
-package src;
+package dz2;
 
 import java.util.Scanner;
