+++ dz2/n_array_of_positive_numbers.java
@@ -1,4 +1,4 @@
-package src;
+package dz2;
 
 import java.util.Scanner;
