+++ dz2/rondom.java
@@ -1,3 +1,5 @@
+package dz2;
+
 import java.util.Random;
 import java.util.Scanner;
