+++ dz2/sum_array_elements.java
@@ -1,4 +1,4 @@
-package src;
+package dz2;
 
 import java.util.Scanner;
